Build system improvements
1. Sign the release apk if configured to do it

By default the build will not sign the release apk, but if you define
the right properties it will.  Create a ~/.gradle/gradle.properties file
with these properties: releaseKeystoreFile, releaseKeystorePassword,
releaseKeyAlias, and releaseKeyPassword.

For example:

releaseKeystoreFile=/Users/tcabot/release-keystore
releaseKeystorePassword=k3yst0r3
releaseKeyAlias=affdexme
releaseKeyPassword=k3y

2. Allow debug builds to coexist with releases

The debug build has a different id (com.affectiva.affdexme.debug) so
android treats it as a completely different app.  By default, though,
both apps look the same on the app launch screen so the debug version is
now called AffdexMeDbg so you can have both at the same time and be able
to tell which is which.

3. Make the apk filename more descriptive

It was just "app-debug.apk" or "app-release.apk".  This commit changes
the "app" to "AffdexMe" and adds the human-readable versionName and
integer versionCode to the filename.

I would have liked the filename to track the cosmetic name (i.e., have
the debug file be called "AffdexMeDbg-....apk") but that turned out to
be surprisingly difficult given the way the build system is set up.  I
think it should be possible but this is the best bang for the buck for
the amount of effort I'm willing to put in at this point.